requirements.txt
matplotlib==3.3.4
easydict==1.9
# opencv_contrib_python
torchvision==0.9.0
mmcv_full==1.2.7
numpy==1.17.0
torch==1.8.0
albumentations==0.5.1
termcolor==1.1.0
attrs==21.2.0
timm==0.4.12
scikit_image==0.17.2
scipy==1.5.4
Cython==0.29.23
tqdm==4.61.0
attr==0.3.1
ipython==7.16.3
mmcv==1.4.1
Pillow==8.4.0
PyYAML==6.0
thop
wandb
